+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Migrated to Rust 1.82.0.
+- `Synchronizer.rewindToNearestHeight` now returns the block height that was
+  actually rewound to, or `null` if no rewind was performed.
 - `Synchronizer.proposeTransfer` throws `TransactionEncoderException.ProposalFromParametersException`
 - `Synchronizer.proposeShielding` throws `TransactionEncoderException.ProposalShieldingException`
 - `Synchronizer.createProposedTransactions` throws `TransactionEncoderException.TransactionNotCreatedException` and `TransactionEncoderException.TransactionNotFoundException`
@@ -22,6 +24,10 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `FailedSynchronizationException` reported using `Synchronizer.onProcessorErrorHandler` now contains the full 
   stacktrace history
+
+### Removed
+- `Synchronizer.getNearestRewindHeight` (its function is now handled internally
+  by `Synchronizer.rewindToNearestHeight`).
 
 ## [2.2.4] - 2024-09-16